Remove unused crate::ws
