+++ src/mux/inner.rs
@@ -6,8 +6,7 @@ use crate::dupe::Dupe;
 use crate::frame::{BindPayload, ConnectPayload, FinalizedFrame, Frame, OpCode, Payload};
 use crate::stream::MuxStream;
 use crate::timing::{OptionalDuration, OptionalInterval};
-use crate::ws::{Message, WebSocketStream};
-use crate::{Datagram, config};
+use crate::{Datagram, Message, WebSocketStream, config};
 use crate::{Error, Result};
 use bytes::Bytes;
 use futures_util::future::poll_fn;
